Change Stage
- 스테이지 전환
- 포탈 생성

+++ Assets/Scripts/Controllers/EnemyController.cs
@@ -8,7 +8,7 @@ using UnityEngine.U2D;
 public class EnemyController : MonoBehaviour
 {
     [SerializeField]
-    int _speed = 5;
+    float _speed = 5;
 
     [SerializeField]
     float CHASE_TIME = 5;
